feedback sw (usize/u64) + nasty merge issue with ComponentEventsDropped

+++ src/sinks/influxdb/metrics.rs
@@ -269,7 +269,7 @@ fn encode_events(
     quantiles: &[f64],
 ) -> BytesMut {
     let mut output = BytesMut::new();
-    let count = events.len() as u64;
+    let count = events.len();
 
     for event in events.into_iter() {
         let fullname = encode_namespace(event.namespace().or(default_namespace), '.', event.name());
